сделал  боковое меню на свои макеты 1 поменял линеа лейауты на констрейн, иначе тулбар закрывал содержимое 2 в манифесте выставил режимы SingleTask для обоих активити, чтобы не выло возвратов по кнопке назад 3 все элементы перевёл на библиотеку androidx 4 в main активити нашел и инициализировал все action элементы
