+++ lesson7_homework/src/main/AndroidManifest.xml
@@ -10,16 +10,14 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".TestActivity"
-            android:theme="@style/AppTheme.NoActionBar" />
-        <activity
             android:name=".DetailActivity"
+            android:launchMode="singleTask"
             android:label="@string/WhetherInCity" />
         <activity
             android:name=".MainActivity"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
-            android:launchMode="singleTop"
+            android:launchMode="singleTask"
             android:theme="@style/ChooseThem">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
