+++ lesson7_homework/src/main/res/layout-land/activity_main.xml
@@ -2,26 +2,26 @@
 
 <!-- это макет для альбомной ориентации здесь два фрагмента-->
 <!--один - статический citiesWhether и другой - динамический whether_in_citys-->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="horizontal"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/content"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:baselineAligned="false">
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:showIn="@layout/app_bar_main">
     <!-- Фрагмент с выбором городов -->
-    <fragment
-        android:name="com.geekbrains.city_weather.frag.ChooseCityFrag"
-        android:id="@+id/citiesWhether"
-        android:layout_weight="1"
-        android:layout_width="0px"
-        android:layout_height="match_parent" />
     <!-- Здесь будет погода выбранного города -->
     <FrameLayout
         android:id="@+id/whether_in_citys"
-        android:layout_width="0px"
+        android:layout_width="0dp"
         android:layout_height="match_parent"
         android:layout_weight="1"
-        android:background="?android:attr/detailsElementBackground">
+        android:background="?android:attr/detailsElementBackground"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/citiesWhether"
+        app:layout_constraintTop_toTopOf="parent">
         <!--это плавающая кнопка для установки города-->
         <com.google.android.material.floatingactionbutton.FloatingActionButton
             android:id="@+id/fab_main"
@@ -34,4 +34,15 @@
             app:srcCompat="@drawable/ic_my_location_white_24dp" />
 
     </FrameLayout>
-</LinearLayout>
+
+    <fragment
+        android:id="@+id/citiesWhether"
+        android:name="com.geekbrains.city_weather.frag.ChooseCityFrag"
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:layout_weight="1"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toStartOf="@+id/whether_in_citys"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+</androidx.constraintlayout.widget.ConstraintLayout>
