+++ lesson7_homework/src/main/res/layout/app_bar_main.xml
@@ -18,14 +18,6 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:srcCompat="@android:drawable/ic_dialog_email" />
-
-    <include layout="@layout/content_main" />
+    <include layout="@layout/activity_main" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
