+++ lesson7_homework/src/main/res/values/styles.xml
@@ -13,7 +13,7 @@
         <item name="android:textColor">@android:color/white</item>
     </style>
 
-    <style name="ChooseThem" parent="AppTheme">
+    <style name="ChooseThem" parent="AppTheme.NoActionBar">
         <item name="colorAccent">#008577</item>
     </style>
